+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Changes
 * [UI/Developer]: Fixed issue where search for member to add to group resulted in no results found when search term contained capital letters.
 * [Documentation]: Added information on fixing ONLY_FULL_GROUP_BY sql error to the administrator faq docs.
 * [UI]: User on a remote project with a project role of manager has the ability to assign user groups to the project.
+* [UI/Developer]: Fixed issue with deleting a user group if it is linked to any projects.
 
 0.22.0 to 19.01
 ----------------